split up service worker helper into tree-shakeable functions
