+++ src/page/utils/OneSignalShimLoader.test.ts
@@ -1,5 +1,5 @@
-import { OneSignalShimLoader } from './OneSignalShimLoader';
 import * as BrowserSupportsPush from './BrowserSupportsPush';
+import { start } from './OneSignalShimLoader';
 
 const supportsSpy = vi
   .spyOn(BrowserSupportsPush, 'isPushNotificationsSupported')
@@ -10,7 +10,7 @@ const isIosSafariSpy = vi
 
 describe('OneSignalShimLoader', () => {
   test('start should load the page script', async () => {
-    OneSignalShimLoader.start();
+    start();
 
     // should load the page script if supported
     const script = document.querySelector('script')!;
@@ -21,10 +21,10 @@ describe('OneSignalShimLoader', () => {
     isIosSafariSpy.mockReturnValue(true);
     vi.spyOn(console, 'info').mockImplementation(() => null);
 
-    OneSignalShimLoader.start();
+    start();
     expect(console.info).toHaveBeenCalledWith(
       expect.stringContaining(
-        'OneSignal: SDK is not compatible with this browser. To support iOS please install as a Web App.',
+        'Incompatible browser. Try these steps: https://tinyurl.com/bdh2j9f7',
       ),
     );
   });
